Make inspect/ClassProps NG-only
