+++ src/attr/__init__.py
@@ -10,11 +10,10 @@ from typing import Callable, Literal, Protocol
 from . import converters, exceptions, filters, setters, validators
 from ._cmp import cmp_using
 from ._config import get_run_validators, set_run_validators
-from ._funcs import asdict, assoc, astuple, has, inspect, resolve_types
+from ._funcs import asdict, assoc, astuple, has, resolve_types
 from ._make import (
     NOTHING,
     Attribute,
-    ClassProps,
     Converter,
     Factory,
     _Nothing,
@@ -45,7 +44,6 @@ __all__ = [
     "NOTHING",
     "Attribute",
     "AttrsInstance",
-    "ClassProps",
     "Converter",
     "Factory",
     "NothingType",
@@ -69,7 +67,6 @@ __all__ = [
     "get_run_validators",
     "has",
     "ib",
-    "inspect",
     "make_class",
     "mutable",
     "resolve_types",
