+++ src/attrs/__init__.py
@@ -4,7 +4,6 @@ from attr import (
     NOTHING,
     Attribute,
     AttrsInstance,
-    ClassProps,
     Converter,
     Factory,
     NothingType,
@@ -18,12 +17,13 @@ from attr import (
     fields_dict,
     frozen,
     has,
-    inspect,
     make_class,
     mutable,
     resolve_types,
     validate,
 )
+from attr._funcs import inspect
+from attr._make import ClassProps
 from attr._next_gen import asdict, astuple
 
 from . import converters, exceptions, filters, setters, validators
